refactor: rename subpackage base_computation into core
