+++ shot_fpfh/descriptors/shot_parallelization.py
@@ -7,7 +7,7 @@ import numpy.typing as npt
 from sklearn.neighbors import KDTree
 from tqdm import tqdm
 
-from shot_fpfh.base_computation import grid_subsampling
+from shot_fpfh.core import grid_subsampling
 
 from .shot import compute_single_shot_descriptor, get_local_rf
